+++ NEWS.md
@@ -1,3 +1,10 @@
+## strsim 1.1.0
+
+- Added neighborhood density and neighbourhood frequency functions. Neighborhood density
+  function is way faster than the previous calculate_coltheartN() function.
+- Removed calculate_coltheartN() and coltheartN() functions.
+- old20 function uses now data.table.
+
 ## strsim 1.0.0
 
-- First release
+- First release.